backend/requirements.txt
django>=4.2
djangorestframework>=3.14
pillow>=10.0
pytesseract>=0.3
python-dotenv>=1.0
requests>=2.31
django-cors-headers>=4.0
